rename: utils 내의 유틸 함수들 lib 내로 옮기기
